pyproject.toml
[tool.poetry]
name = "graf-zeppelin"
version = "0.1.0"
description = ""
authors = ["Lunarmagpie <Bambolambo0@gmail.com>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
hikari = "^2.0.0-alpha.106"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
flake8 = "^4.0.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
